requirements.txt
-i https://pypi.org/simple
apscheduler==3.6.3
attrs==19.3.0
awsebcli==3.18.1
bcrypt==3.1.7
blessed==1.17.5
botocore==1.15.49
cached-property==1.5.1
cement==2.8.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==2.9.2
docker-compose==1.25.5
docker[ssh]==4.2.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
feedparser==5.2.1
flask==1.1.2
future==0.16.0
idna==2.7
itsdangerous==1.1.0
jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
markupsafe==1.1.1
paramiko==2.7.1
pathspec==0.5.9
pg8000==1.15.2
pycparser==2.20
pygresql==5.1.2
pynacl==1.3.0
pyrsistent==0.16.0
python-dateutil==2.8.0
pytz==2020.1
pyyaml==5.3.1
requests==2.20.1
scramp==1.1.1
semantic-version==2.5.0
six==1.11.0
sqlalchemy==1.3.17
termcolor==1.1.0
texttable==1.6.2
tzlocal==2.1
urllib3==1.24.3
wcwidth==0.1.9
websocket-client==0.57.0
werkzeug==1.0.1
